+++ src/compiler/llvm_codegen_stmt.c
@@ -199,8 +199,9 @@ static void llvm_emit_cond(GenContext *c, BEValue *be_value, Expr *expr, bool bo
 	if (bool_cast && be_value->type != type_bool)
 	{
 		Type *type = be_value->type;
-		CastKind cast = cast_to_bool_kind(type);
-		llvm_emit_cast(c, cast, last, be_value, type_bool, type);
+		llvm_emit_bool_cast(c, last, be_value);
+
+//		llvm_emit_cast(c, cast, last, be_value, type_bool, type);
 	}
 }