+++ src/compiler/parse_global.c
@@ -2140,6 +2140,10 @@ static inline bool parse_func_macro_header(ParseContext *c, Decl *decl)
 		// 5d. A method type but no dot is also wrong.
 		if (method_type->kind == TYPE_INFO_IDENTIFIER && tok_is(c, TOKEN_LPAREN))
 		{
+			if (method_type->resolve_status == RESOLVE_DONE && !type_is_user_defined(method_type->type))
+			{
+				RETURN_PRINT_ERROR_AT(false, method_type, "A function name may not use a reserved type name like '%s'.", method_type->type->name);
+			}
 			RETURN_PRINT_ERROR_AT(false, method_type, "A function name must start with lower case, as names starting with upper case are reserved for types.");
 		}
 		RETURN_PRINT_ERROR_AT(false, method_type, "There is unexpectedly a type after the return type, did you forget a '.'?");